+++ src/main/java/com/example/exchange/dto/response/CurrencyRateResponseDto.java
@@ -1,11 +1,13 @@
 package com.example.exchange.dto.response;
 
+import java.math.BigDecimal;
+
 public class CurrencyRateResponseDto {
     public String currencyFrom;
     public String currencyTo;
-    public float rate;
+    public BigDecimal rate;
 
-    public CurrencyRateResponseDto(String currencyFrom, String currencyTo, float rate) {
+    public CurrencyRateResponseDto(String currencyFrom, String currencyTo, BigDecimal rate) {
         this.currencyFrom = currencyFrom;
         this.currencyTo = currencyTo;
         this.rate = rate;
